Fix dtype mismatch in MoE _dispatch index_put
The EDM sampler runs with float64 latents while the expert modules return
float32 under autocast. The non-MoE path returned the float32 tensor
directly and relied on implicit promotion in the sampler's arithmetic, but
_dispatch assigns via index_put (out[sel] = ...), which requires an exact
dtype match and raised 'Index put requires the source and destination
dtypes match, got Double for the destination and Float for the source'.

Cast the module output to out's dtype. Add a regression test with float64
input and float32-returning experts.

+++ fme/downscaling/distillation/fastgen_teacher.py
@@ -358,7 +358,11 @@ class AceDiffusionTeacher(FastGenNetwork):
             # Pass the scalar t directly when broadcasting (it broadcasts over
             # the selected subset); otherwise index t to preserve its shape.
             sigma_sel = t if broadcast else t[sel]
-            out[sel] = module(x_t[sel], cond_sel, sigma_sel)
+            # Cast to out's dtype: the EDM sampler's latents are float64 while
+            # the module returns float32 (autocast), and index_put requires an
+            # exact dtype match (unlike the arithmetic the non-MoE path relied
+            # on, which promotes implicitly).
+            out[sel] = module(x_t[sel], cond_sel, sigma_sel).to(out.dtype)
 
         for (lo, hi), module in zip(self._moe_sigma_ranges, self._moe_experts):
             sel = remaining & (route >= lo) & (route <= hi)
